Style: post 폴더구조 변경 #24
