+++ src/pages/MyPage/Posts/index.jsx
@@ -1,10 +1,7 @@
 import { useEffect, useState } from 'react';
 import * as s from './style';
-// import dotdotdot from "../../../assets/dotdotdot.svg";
-// import UnLikedIcon from "../../../assets/UnLikedIcon.svg";
-import MyPosts from '../../../Temp/mypage/mypost/MyPosts';
-import CreatePost from './ArtistPosts';
-
+import dotdotdot from '../../../assets/dotdotdot.svg';
+import UnLikedIcon from '../../../assets/UnLikedIcon.svg';
 import mypageAPI from '../../../api/mypage/mypageAPI';
 // import LikedIcon from "../../../assets/LikedIcon.svg";
 
@@ -12,15 +9,39 @@ const Posts = () => {
   const [userId, setUserId] = useState(1);
   const [postArtistList, setPostArtistList] = useState([]);
   const [selectedArtist, setSelectedArtist] = useState();
-
-  const onClickArtist = (artistId) => {
-    setSelectedArtist(artistId);
-  };
+  const [artistPost, setArtistPost] = useState([]);
+  const [isClickDot, setIsClickDot] = useState(false);
 
   const getMyPostArtistTab = () => {
     mypageAPI.getMyPostArtistTab(userId).then((data) => {
       setPostArtistList(data.postArtistList);
       setSelectedArtist(data.postArtistList[0].artistId);
+    });
+  };
+
+  const getMyPost = () => {
+    mypageAPI.getMyPost(userId, selectedArtist).then((data) => {
+      setArtistPost(data.myPostList);
+    });
+  };
+
+  const onClickArtist = (artistId) => {
+    setSelectedArtist(artistId);
+  };
+
+  const handleClickdot = () => {
+    setIsClickDot((prev) => !prev);
+  };
+
+  const handleClickDel = (userId, postId) => {
+    setIsClickDot(false);
+    deleteMyPost(userId, postId);
+    getMyPost(userId, selectedArtist);
+  };
+
+  const deleteMyPost = (postId) => {
+    mypageAPI.deleteMyPost(userId, postId).then((data) => {
+      console.log(data);
     });
   };
 
@@ -36,77 +57,89 @@ const Posts = () => {
     );
   });
 
+  //포스트 목록
+  const ArtistPosts = () => {
+    return (
+      <>
+        {selectedArtist === 0 ? (
+          <div
+            style={{
+              color: 'gray',
+              fontWeight: 'bold',
+              width: '100%',
+              height: '200px',
+              display: 'flex',
+              flexDirection: 'column',
+              justifyContent: 'center',
+              alignItems: 'center',
+            }}
+          >
+            아티스트의 도안을 꾸미고 포스트해보세요!
+          </div>
+        ) : (
+          artistPost?.map((item) => (
+            <>
+              <s.PostFrame key={`post_${item.postId}`}>
+                <s.leftContainer>
+                  <s.PostImageFrame>
+                    <s.PostImage src={item.polaroid} />
+                  </s.PostImageFrame>
+                </s.leftContainer>
+                <s.rightContainer>
+                  <s.postInfoWrapper>
+                    <s.nicknameWrapper>{item.nickname}</s.nicknameWrapper>
+                    <s.tagsWrapper>
+                      #{item.enterComp} #{item.groupName} <br />
+                      {item.groupName === item.memberName
+                        ? ''
+                        : `#${item.memberName}`}
+                      #{item.albumName}
+                    </s.tagsWrapper>
+                    <s.dateWrapper>
+                      {item?.postDateTime.slice(
+                        0,
+                        item?.postDateTime.indexOf('T')
+                      )}
+                    </s.dateWrapper>
+                    <s.likeWrapper>
+                      <s.likeIcon src={UnLikedIcon} />
+                      <s.likeCount key={item.id}>{item.likeQuant}</s.likeCount>
+                    </s.likeWrapper>
+                  </s.postInfoWrapper>
+                  <s.moreIconWrapper>
+                    <s.moreIcon src={dotdotdot} onClick={handleClickdot} />
+                  </s.moreIconWrapper>
+                </s.rightContainer>
+              </s.PostFrame>
+              {isClickDot && (
+                <s.deleteButton onClick={() => handleClickDel(item.postId)}>
+                  삭제
+                </s.deleteButton>
+              )}
+            </>
+          ))
+        )}
+      </>
+    );
+  };
+
   useEffect(() => {
     getMyPostArtistTab();
   }, []);
+  useEffect(() => {
+    getMyPost();
+  }, [selectedArtist]);
 
   return (
     <>
       <s.Wrapper>
         <s.ArtistsTabWrapper>{artists}</s.ArtistsTabWrapper>
         <s.PostsWrapper>
-          <CreatePost
-            selectedArtist={selectedArtist}
-            // setCurrentArtist={setCurrentArtist}
-          />
-          {/* {posts.filter(post => post.groupName === selectedArtist)} */}
+          <ArtistPosts />
         </s.PostsWrapper>
       </s.Wrapper>
     </>
   );
 };
-
-// const CreatePost = ({ selectedArtistInfo, setCurrentArtist, post }) => {
-
-//   const artistPosts = selectedArtistInfo ? selectedArtistInfo.posts : [];
-
-//   return (
-//     <>
-//     <s.PostFrame>
-//       {artistPosts.length === 0 ? (
-//         <div>아티스트의 도안을 꾸미고 포스트해보세요!</div>
-//       ) : (
-//         artistPosts.map((item, index) => (
-//           <>
-//           <s.leftContainer>
-//             <s.PostImageFrame >
-//               <s.PostImage src={item.fileUrl} />
-//             </s.PostImageFrame>
-//           </s.leftContainer>
-//             <s.rightContainer>
-//               <s.postInfoWrapper
-//                 key={item.id}
-//               >
-//                 <s.nicknameWrapper>OnPol1004</s.nicknameWrapper>
-//                 <s.tagsWrapper>
-//                   #소속사 #아티스트 <br/> #멤버이름 #컬렉션명
-//                 </s.tagsWrapper>
-//                 <s.dateWrapper>
-//                   2023.10.13
-//                 </s.dateWrapper>
-//                 <s.likeWrapper>
-//                   <s.likeIcon
-//                     src={ UnLikedIcon }
-//                   />
-//                   <s.likeCount
-//                     key={item.id}
-//                   >
-//                     {item.likeCount}
-//                   </s.likeCount>
-//                 </s.likeWrapper>
-//               </s.postInfoWrapper>
-//               <s.moreIconWrapper>
-//                 <s.moreIcon
-//                   src={dotdotdot}
-//                 />
-//               </s.moreIconWrapper>
-//             </s.rightContainer>
-//             </>
-//           ))
-//         )}
-//       </s.PostFrame>
-//     </>
-//   )
-// };
 
 export default Posts;
